Add AMD specific hardware acceleration option to match NVIDIA and Intel
Mainly adds drivers to the Dockerfile, but copies the intel directory layout, and adds to the build script. The .docker/base/amd/ folder isn't strictly needed. It could point to intel instead, the files are the same. But I did this rather than renaming the directory to intel-amd or vaapi